push inte ttest
